Add typescript - use custom hook for metamask
